third_party/blink/web_tests/external/wpt/css/css-ui/user-select-inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] user-select:all should not be inherited.
  assert_equals: user-select:all should not be inherited. expected "auto" but got "all"
[FAIL] user-select:none should not be inherited.
  assert_equals: user-select:none should not be inherited. expected "auto" but got "none"
[FAIL] user-select:text should not be inherited.
  assert_equals: user-select:text should not be inherited. expected "auto" but got "text"
Harness: the test ran to completion.
